add ruby codegen (#22)
## Description
To add Ruby codegen for UPI to support fluentd parser which is written
in Ruby.

Buf is upgraded and some of the yaml syntax are changed. Custom remote
plugins will [deprecated](https://buf.build/grpc/plugins). The plugin
that can be ported are following the yaml syntax in the [curated
plugins](https://buf.build/plugins).

Python plugin are not moved as the
[dependencies](https://buf.build/jeffsawatzky/plugins/mypy-protobuf) are
not migrated yet

## Key Changes
- `buf.gen.yaml` - added Ruby codegen plugin, modified yaml syntax for
Go plugin
- `.github/workflows/ci.yaml` - upgrade buf to v1.11
- `.github/workflows/publish-ruby.yaml` - to publish proto to ruby gem
(version will take from gemspec)

`caraml-upi-protos` v0.0.0 was published manually as part of testing.
Before show CI trying to re-push (and expected rejection)

![image](https://user-images.githubusercontent.com/30390872/212017585-efed74e2-287b-46d4-9f28-03e579787745.png)

Co-authored-by: ningjie.lee <ningjie.lee@gojek.com>